# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-checklist
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /employee-checklists/{employeeChecklistId}/email:
    post:
      tags:
      - Communication resources
      summary: Send email notification
      description: Send an email notification to the manager for the employee checklist
        that matches provided id
      operationId: sendEmail
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Successful Operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /checklists:
    get:
      tags:
      - Checklist resources
      summary: Get all checklists
      operationId: fetchAllChecklists
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checklist"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Checklist resources
      summary: Create a new checklist
      operationId: createChecklist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistCreateRequest"
        required: true
      responses:
        "201":
          description: Successful Operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /checklists/{checklistId}:
    get:
      tags:
      - Checklist resources
      summary: Fetch checklist by id
      operationId: fetchChecklistById
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Checklist resources
      summary: Delete a checklist
      operationId: deleteChecklist
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Checklist resources
      summary: Update a checklist
      operationId: updateChecklist
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /checklists/{checklistId}/activate:
    patch:
      tags:
      - Checklist resources
      summary: Activate checklist
      operationId: activateChecklist
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /employee-checklists/{employeeChecklistId}/correspondence:
    get:
      tags:
      - Communication resources
      summary: Fetch correspondence
      description: Fetch the correspondence that has occured for an employee checklist
      operationId: fetchCorrespondence
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Correspondence"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Correspondence"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    CommunicationChannel:
      type: string
      description: Valid channels to use when communicating with the organization
      writeOnly: true
      enum:
      - EMAIL
      - NO_COMMUNICATION
...
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    QuestionType:
      type: string
      description: The question type of the task
      writeOnly: true
      enum:
      - YES_OR_NO
      - YES_OR_NO_WITH_TEXT
      - COMPLETED_OR_NOT_RELEVANT
      - COMPLETED_OR_NOT_RELEVANT_WITH_TEXT
...
    RoleType:
      type: string
      description: The role type of the task
      readOnly: true
      enum:
      - EMPLOYEE
      - MANAGER
...
    ChecklistCreateRequest:
      required:
      - displayName
      - name
      - organizationNumber
      - roleType
      type: object
      properties:
        name:
          type: string
          description: The name of the checklist
          example: Checklist_A
        displayName:
          type: string
          description: The display name of the checklist
          example: Display name
        organizationNumber:
          type: integer
          description: The organization that the checklist is created for
          format: int32
          example: 11
        roleType:
          $ref: "#/components/schemas/RoleType"
      description: Model for checklist create request
    Permission:
      type: string
      description: The permission needed to administrate the phase
      writeOnly: true
      enum:
      - SUPERADMIN
      - ADMIN
...
    Checklist:
      type: object
      properties:
        id:
          type: string
          description: The id of the checklist
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        roleType:
          $ref: "#/components/schemas/RoleType"
        name:
          type: string
          description: The name of the checklist
          readOnly: true
          example: Checklist_A
        displayName:
          type: string
          description: The display name of the checklist
          readOnly: true
          example: Display name
        version:
          type: integer
          description: The version of the checklist
          format: int32
          readOnly: true
          example: 1
        lifeCycle:
          $ref: "#/components/schemas/LifeCycle"
        created:
          type: string
          description: The created date and time of the checklist
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The last update date and time of the checklist
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        phases:
          type: array
          description: Phases in the checklist
          readOnly: true
          items:
            $ref: "#/components/schemas/Phase"
      description: Model for checklist
      readOnly: true
    LifeCycle:
      type: string
      description: The lifecycle of the checklist
      readOnly: true
      enum:
      - CREATED
      - ACTIVE
      - DEPRECATED
      - RETIRED
...
    Phase:
      type: object
      properties:
        id:
          type: string
          description: The id of the phase
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        name:
          type: string
          description: The name of the phase
          readOnly: true
          example: Första veckan
        bodyText:
          type: string
          description: The body text of the phase
          readOnly: true
          example: Detta är en beskrivning av vad som ska göras under första veckan
        timeToComplete:
          type: string
          description: The time to complete the phase
          readOnly: true
          example: P1M
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        sortOrder:
          type: integer
          description: The sort order of the phase
          format: int32
          readOnly: true
          example: 1
        tasks:
          type: array
          description: Tasks in the phase
          readOnly: true
          items:
            $ref: "#/components/schemas/Task"
        created:
          type: string
          description: The created date and time of the phase
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The last update date and time of the phase
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Model for a phase
    Task:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order of the task
          format: int32
          readOnly: true
          example: 1
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        permission:
          $ref: "#/components/schemas/Permission"
        created:
          type: string
          description: The date and time the task was created
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Task model
...
    ChecklistUpdateRequest:
      type: object
      properties:
        displayName:
          type: string
          description: The name of the checklist
          example: New display name
        roleType:
          $ref: "#/components/schemas/RoleType"
      description: Model for checklist update request
...
    Correspondence:
      type: object
      properties:
        messageId:
          type: string
          description: Id for message
          readOnly: true
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        recipient:
          type: string
          description: Message recipient
          readOnly: true
          example: email.address@noreply.com
        attempts:
          type: integer
          description: Attempt count
          format: int32
          readOnly: true
          example: 1
        correspondenceStatus:
          $ref: "#/components/schemas/CorrespondenceStatus"
        communicationChannel:
          $ref: "#/components/schemas/CommunicationChannel"
        sent:
          type: string
          description: Timestamp when message was sent
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+02:00
      description: Model for correspondence
    CorrespondenceStatus:
      type: string
      description: Status for correspondence
      readOnly: true
      enum:
      - SENT
      - NOT_SENT
      - ERROR
      - WILL_NOT_SEND
...
  securitySchemes: {}